Move Sufia::User into Curate::UserBehavior
With Sufia included in Curate, we can now combine the two modules.
